Modified sender_cadmium.hpp acc. to coding conventions
